Stress development maturity updates (#3481)
* Support local addons path override in stress test deployment

* Support username based deployId in local stress deployment

* Fix network chaos example

* Add stress test example release ci

* Support WhatIf in stress infrastructure provision script

* Fix image ref issues in powershell template

* Add stress cluster provision and validation pipeline

* Simplify provision pipeline yaml and enable sub config/sp auth

* Fix chart.lock local reference in stress example

* Simplify stress user detection


* Run helm plugin add with helper

* Use mms pool for stress test release

* Fix default parameter set for stress cluster provision script

* Add WhatIf support to ps module install helper function

+++ tools/stress-cluster/cluster/provision.ps1
@@ -1,32 +1,80 @@
+[CmdletBinding(DefaultParameterSetName = 'Default', SupportsShouldProcess = $true)]
 param (
     [string]$Environment = 'dev',
     [string]$Namespace = 'stress-infra',
-    [switch]$Development = $false
+    [switch]$Development = $false,
+
+    [Parameter(ParameterSetName = 'Provisioner', Mandatory = $true)]
+    [ValidateNotNullOrEmpty()]
+    [string] $TenantId,
+
+    [Parameter(ParameterSetName = 'Provisioner', Mandatory = $true)]
+    [ValidatePattern('^[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$')]
+    [string] $ProvisionerApplicationId,
+
+    [Parameter(ParameterSetName = 'Provisioner', Mandatory = $true)]
+    [string] $ProvisionerApplicationSecret,
+
+    # Enables passing full json credential config without throwing unrecognized parameter errors
+    [Parameter(ValueFromRemainingArguments = $true)]
+    $RemainingArguments
 )
+
+$ErrorActionPreference = 'Stop'
+. (Join-Path $PSScriptRoot "../../../eng/common/scripts/Helpers" PSModule-Helpers.ps1)
+Install-ModuleIfNotInstalled -WhatIf:$false "powershell-yaml" "0.4.1" | Import-Module
 
 $STATIC_TEST_DOTENV_NAME="public"
 $VALUES_FILE = "$PSScriptRoot/kubernetes/stress-test-addons/values.yaml"
 $STRESS_CLUSTER_RESOURCE_GROUP = ""
 
-function Run()
+function _run([string]$CustomWhatIfFlag)
 {
-    Write-Host "`n==> $args`n" -ForegroundColor Green
-    $command, $arguments = $args
-    & $command $arguments
+    if ($WhatIfPreference -and [string]::IsNullOrEmpty($CustomWhatIfFlag)) {
+        Write-Host "`n==> [What if] $args`n" -ForegroundColor Green
+        return
+    } else {
+        $cmdArgs = $args
+        if ($WhatIfPreference) {
+            $cmdArgs += $CustomWhatIfFlag
+        }
+        Write-Host "`n==> $cmdArgs`n" -ForegroundColor Green
+        $command, $arguments = $cmdArgs
+        & $command $arguments
+    }
     if ($LASTEXITCODE) {
         Write-Error "Command '$args' failed with code: $LASTEXITCODE" -ErrorAction 'Continue'
     }
 }
 
-function RunOrExitOnFailure()
+function Run()
 {
-    Run @args
+    _run '' @args
+}
+
+function RunSupportingWhatIfFlag([string]$CustomWhatIfFlag)
+{
+    if ($WhatIfPreference) {
+        _run $CustomWhatIfFlag @args
+    } else {
+        _run @args
+    }
     if ($LASTEXITCODE) {
         exit $LASTEXITCODE
     }
 }
 
-function DeployStaticResources([hashtable]$params) {
+function RunOrExitOnFailure()
+{
+    $LASTEXITCODE = 0
+    _run '' @args
+    if ($LASTEXITCODE) {
+        exit $LASTEXITCODE
+    }
+}
+
+function DeployStaticResources([hashtable]$params)
+{
     Write-Host "Deploying static resources"
 
     RunOrExitOnFailure az group create `
@@ -82,7 +130,9 @@ function DeployStaticResources([hashtable]$params) {
     $dotenv = $credentials.GetEnumerator() | ForEach-Object { "$($_.Key)=$($_.Value)`n" }
     (-join $dotenv) | Out-File $envFile
     Run az keyvault secret set --vault-name $params.staticTestSecretsKeyvaultName --file $envFile -n $STATIC_TEST_DOTENV_NAME
-    Remove-Item -Force $envFile
+    if (Test-Path $envFile) {
+        Remove-Item -Force $envFile
+    }
     if ($LASTEXITCODE) {
         exit $LASTEXITCODE
     }
@@ -90,24 +140,28 @@ function DeployStaticResources([hashtable]$params) {
     SetEnvProvisioner $spInfo
 }
 
-function GetEnvValues() {
+function GetEnvValues()
+{
     $values = ConvertFrom-Yaml -Ordered (Get-Content -Raw $VALUES_FILE)
     return $values
 }
 
-function SetEnvValues([object]$values) {
+function SetEnvValues([object]$values)
+{
     $values | ConvertTo-Yaml | Out-File $VALUES_FILE
     Write-Warning "Update https://aka.ms/azsdk/stress/dashboard link page with new dashboard link: $($outputs.DASHBOARD_LINK.value)"
     Write-Warning "$VALUES_FILE has been updated and must be checked in."
 }
 
-function SetEnvProvisioner([object]$provisioner) {
+function SetEnvProvisioner([object]$provisioner)
+{
     $values = GetEnvValues
     $values.provisionerAppId.$Environment = $provisioner.appId
     SetEnvValues $values
 }
 
-function SetEnvOutputs([hashtable]$params) {
+function SetEnvOutputs([hashtable]$params)
+{
     $outputs = (az deployment sub show `
         -o json `
         -n "stress-deploy-$($params.groupSuffix)" `
@@ -131,9 +185,10 @@ function SetEnvOutputs([hashtable]$params) {
     SetEnvValues $values
 }
 
-function DeployClusterResources([hashtable]$params) {
+function DeployClusterResources([hashtable]$params)
+{
     Write-Host "Deploying stress cluster resources"
-    RunOrExitOnFailure az deployment sub create `
+    RunSupportingWhatIfFlag "--what-if" az deployment sub create `
         -o json `
         --subscription $params.subscriptionId `
         -n "stress-deploy-$($params.groupSuffix)" `
@@ -152,23 +207,30 @@ function DeployClusterResources([hashtable]$params) {
         --subscription $params.subscriptionId
 }
 
-function DeployHelmResources() {
+function DeployHelmResources()
+{
     Write-Host "Installing stress infrastructure charts"
+    $LastWhatIfPreference = $WhatIfPreference
+
+    $WhatIfPreference = $false
     RunOrExitOnFailure helm repo add chaos-mesh https://charts.chaos-mesh.org
     RunOrExitOnFailure helm dependency update $PSScriptRoot/kubernetes/stress-infrastructure
-    RunOrExitOnFailure kubectl create namespace $Namespace --dry-run=client -o yaml | kubectl apply -f -
+
+    $WhatIfPreference = $LastWhatIfPreference
+    Run kubectl create namespace $Namespace --dry-run=client -o yaml | kubectl apply -f -
 
     # Skip installing chaos mesh charts in development mode (i.e. when testing stress watcher only).
     $deployChaosMesh = "$(!$Development)".ToLower()
 
-    RunOrExitOnFailure helm upgrade --install stress-infra `
+    RunSupportingWhatIfFlag "--dry-run" helm upgrade --install stress-infra `
         -n $Namespace `
         --set stress-test-addons.env=$Environment `
         --set deploy.chaosmesh=$deployChaosMesh `
         $PSScriptRoot/kubernetes/stress-infrastructure
 }
 
-function LoadEnvParams() {
+function LoadEnvParams()
+{
     try {
         $params = (Get-Content $PSScriptRoot/azure/parameters/$Environment.json | ConvertFrom-Json -AsHashtable).parameters
     } catch {
@@ -184,9 +246,22 @@ function LoadEnvParams() {
     return $paramHash
 }
 
-function main() {
-    # . (Join-Path $PSScriptRoot "../Helpers" PSModule-Helpers.ps1)
-    # Install-ModuleIfNotInstalled "powershell-yaml" "0.4.1" | Import-Module
+function main()
+{
+    # Force a reset of $LASTEXITCODE 0 so that when running in -WhatIf mode
+    # we don't inadvertently check a $LASTEXITCODE value from before the script invocation
+    if ($WhatIfPreference) {
+        helm -h > $null
+    }
+
+    if ($PSCmdlet.ParameterSetName -eq 'Provisioner') {
+        az login `
+            --service-principal `
+            --username $ProvisionerApplicationId `
+            --password $ProvisionerApplicationSecret`
+            --tenant $TenantId
+        if ($LASTEXITCODE) { exit $LASTEXITCODE }
+    }
 
     if (!$Development) {
         $params = LoadEnvParams
@@ -197,8 +272,4 @@ function main() {
     DeployHelmResources
 }
 
-# Don't call functions when the script is being dot sourced
-if ($MyInvocation.InvocationName -ne ".") {
-    $ErrorActionPreference = 'Stop'
-    main
-}
+main
